feat(java): bind scope.failed and read line 0.2
Location lines and columns hold the schema's full range, diagnostics quote
producer values bounded and escaped, and the corpus tests check exact
schema pointers on the per-type branch and that codec pointers refine
them. Maven output under consumer fixtures is ignored.

+++ java/protocol/src/main/java/io/github/yannisyoussef/qe/report/protocol/Location.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 import org.jspecify.annotations.Nullable;
 
 /** Where a test or step is defined. Display only; never resolved on a server. */
-public record Location(String file, @Nullable Integer line, @Nullable Integer column) {
+public record Location(String file, @Nullable Long line, @Nullable Long column) {
   public Location {
     Objects.requireNonNull(file, "file");
   }
